Init secure index page
Add in the secure index page for when a user is authorized throught the django system.

This is the main page from which all other apps will be branched of off. Currently we have two main pages, the secure for all protected apps, and the public for anything that is publiclly available.